requirements.txt

chess==1.11.0
streamlit>=1.37.0
pandas>=2.2.0
Pillow>=10.2.0
cairosvg>=2.7.1
matplotlib>=3.8.0
requests>=2.31.0
